+++ tests/lib/hardware/openstack_libcloud.py
@@ -25,6 +25,7 @@
 import logging
 import threading
 import time
+import subprocess
 
 import libcloud.security
 from libcloud.compute.types import Provider, NodeState, StorageVolumeState
@@ -345,3 +346,27 @@ class Hardware(HardwareBase):
         super().destroy()
         self.conn.ex_delete_security_group(self._ex_security_group)
         self.conn.delete_key_pair(self._ex_os_key)
+
+    def remove_host_keys(self):
+        # The mitogen plugin does not correctly ignore host key checking, so we
+        # should remove any host keys for our nodes before starting.
+        # The 'ssh' connection imports ssh-keys for us, so as a first step we
+        # run a standard ssh connection to do the imports. We could import the
+        # sshkeys manually first, but we also want to wait on the connection to
+        # be available (in order to even be able to get them).
+        # Therefore simply remove any entries from your known_hosts. It's also
+        # helpful to do this after a build to clean up anything locally.
+        for node in self.nodes.values():
+            self.remove_ssh_key(node.get_ssh_ip())
+
+    def remove_ssh_key(self, ip):
+        subprocess.run(
+            "ssh-keygen -R %s" % ip,
+            shell=True
+        )
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.destroy()
