+++ tox.ini
@@ -28,6 +28,8 @@ passenv =
   K8S_VERSION
   KUBERNETES_DEPLOYMENT
   NODE_IMAGE
+  NODE_IMAGE_ID
+  NODE_IMAGE_USER
   NODE_SIZE
   OS_AUTH_URL
   OS_AUTH_VERSION
